+++ servicio-inventario/src/main/java/com/isam/dto/inventario/AjusteLoteDto.java
@@ -1,10 +1,12 @@
 package com.isam.dto.inventario;
 
+import java.math.BigDecimal;
+
 public record AjusteLoteDto(
     String idLote,
     String numeroLote,
     String ubicacion,  // "ALMACEN" o "ESTANTERIA"
-    double cantidadAjustada,  // positivo = añadido, negativo = deducido
-    double stockAnterior,
-    double stockNuevo
+    BigDecimal cantidadAjustada,  // positivo = añadido, negativo = deducido
+    BigDecimal stockAnterior,
+    BigDecimal stockNuevo
 ) {}
