+++ servicio-inventario/src/main/java/com/isam/dto/inventario/InventarioDto.java
@@ -1,12 +1,13 @@
 package com.isam.dto.inventario;
 
+import java.math.BigDecimal;
 
 public record InventarioDto(
     String idInventario,
     String sku,
     String ean,
     String plu,
-    Double cantidadAlmacen,
-    Double cantidadEstanteria,
+    BigDecimal cantidadAlmacen,
+    BigDecimal cantidadEstanteria,
     String unidadMedida
 ) {}
