+++ servicio-inventario/src/main/java/com/isam/dto/inventario/ReporteDiscrepanciasDto.java
@@ -1,14 +1,15 @@
 package com.isam.dto.inventario;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record ReporteDiscrepanciasDto(
     String sku,
-    double stockLogicoEstanteria,
-    double stockFisicoEstanteria,
-    double discrepanciaEstanteria,  // físico - lógico
-    double stockLogicoAlmacen,
-    double stockFisicoAlmacen,
-    double discrepanciaAlmacen,  // físico - lógico
+    BigDecimal stockLogicoEstanteria,
+    BigDecimal stockFisicoEstanteria,
+    BigDecimal discrepanciaEstanteria,  // físico - lógico
+    BigDecimal stockLogicoAlmacen,
+    BigDecimal stockFisicoAlmacen,
+    BigDecimal discrepanciaAlmacen,  // físico - lógico
     List<AjusteLoteDto> ajustesRealizados
 ) {}
