+++ servicio-inventario/src/main/java/com/isam/dto/movimiento/MovimientoInventarioDto.java
@@ -1,11 +1,13 @@
 package com.isam.dto.movimiento;
 
+import java.math.BigDecimal;
+
 public record MovimientoInventarioDto(
     String idMovimiento,
     String sku,
     String idLote,
     String tipoMovimiento,
-    Double cantidad,
+    BigDecimal cantidad,
     String unidadMedida,
     String fechaHora,
     String idUsuario,
